abstract riff state
