+++ packages/media-parser/src/state/riff.ts
@@ -1,0 +1,30 @@
+import type {SpsAndPps} from './parser-state';
+
+type AvcProfileInfoCallback = (profile: SpsAndPps) => Promise<void>;
+
+export const riffSpecificState = () => {
+	let avcProfile: SpsAndPps | null = null;
+
+	const profileCallbacks: AvcProfileInfoCallback[] = [];
+
+	const registerOnAvcProfileCallback = (callback: AvcProfileInfoCallback) => {
+		profileCallbacks.push(callback);
+	};
+
+	const onProfile = async (profile: SpsAndPps) => {
+		avcProfile = profile;
+		for (const callback of profileCallbacks) {
+			await callback(profile);
+		}
+
+		profileCallbacks.length = 0;
+	};
+
+	return {
+		getAvcProfile: () => {
+			return avcProfile;
+		},
+		onProfile,
+		registerOnAvcProfileCallback,
+	};
+};
